refactor: rename context_impl to root_module_impl
